Implementation(?) of A Complete Example from W3C site: https://www.w3.org/TR/annotation-model/#complete-example
Aim of this is to present a demonstration on using Anno4J library and increase familiarity.
There were a number of hiccups along the way: we can reliably create selectors, but how do we
add them to targets, as opposed to annotation itself? Further, how do we manage id's? One
would likely to suggest they are done programmatically, but there is little to no documentation
on this. All in all, a good first step.