Add annotatation Postgre text field
Annotations are now added but there are some problems. It does not
show the newly added annotations now.

Image annotations are mostly complete. Still all this annotations are
dependent on the python server.

Postgresql requires text field for the body of the memory. It has
been fixed with a docarator.